requirements.txt
# Файл списка зависимостей
# Команда для установки пакетов из этого файла:
# pip install -r requirements.txt
# Команда для установки с обновлением:
# pip install --upgrade -r requirements.txt
# Альтернативная команда установки пакетов:
# pip install playwright pytest pytest-html allure-pytest requests locust faker
# Команда формирования этого документа (заменяет старый т.е. этот):
# pip freeze > requirements.txt

# базовые библиотеки тестирования
playwright==1.51.0 # основной фреймворк тестирования ui
pytest==8.3.5 # библиотека управления тестами

# отчёты
pytest-html==4.1.1 # отчеты пайтест в формате хтмл
allure-pytest==2.14.0 # отчеты аллюр и пайтест

# апи
requests==2.32.3 # хттп запросы

# нагрузочное
locust==2.35.0 # простое нагрузочное тестирование

# прочее
allure-python-commons==2.14.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
Faker==37.3.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==25.4.1
geventhttpclient==2.3.3
greenlet==3.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==25.0
pluggy==1.5.0
psutil==7.0.0
pycparser==2.22
pyee==12.1.1
pytest-base-url==2.1.0
pytest-metadata==3.1.1
python-slugify==8.0.4
pywin32==310
pyzmq==26.4.0
setuptools==79.0.0
text-unidecode==1.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2